Extract TypeScript utilities into dedicated module
- Created new TypeScriptUtils module for better code organization
- Removed old AST.ts file and consolidated utilities
- Updated all imports to use the new module structure
- Enhanced modularity using Nano dependency injection pattern
