--- examples/refactors/asyncAwaitToGen_reExport.ts
@@ -1,6 +1,0 @@
-// 4:28
-import { Eff } from "@/utils/reExport.js"
-
-export async function refactorMe(arg: string) {
-  return await Promise.resolve(1)
-}
