--- examples/utils/reExport.ts
@@ -1,1 +1,0 @@
-export * as Eff from "effect/Effect"
