+++ src/completions/contextSelfInClasses.ts
@@ -1,32 +1,24 @@
-import * as Option from "effect/Option"
-import * as AST from "../core/AST"
 import * as LSP from "../core/LSP"
 import * as Nano from "../core/Nano"
 import * as TypeScriptApi from "../core/TypeScriptApi"
+import * as TypeScriptUtils from "../core/TypeScriptUtils"
 
 export const contextSelfInClasses = LSP.createCompletion({
   name: "contextSelfInClasses",
   apply: Nano.fn("contextSelfInClasses")(function*(sourceFile, position) {
     const ts = yield* Nano.service(TypeScriptApi.TypeScriptApi)
+    const tsUtils = yield* Nano.service(TypeScriptUtils.TypeScriptUtils)
 
-    const maybeInfos = yield* Nano.option(
-      AST.parseDataForExtendsClassCompletion(sourceFile, position)
-    )
-    if (Option.isNone(maybeInfos)) return []
-    const { accessedObject, className, replacementSpan } = maybeInfos.value
+    const maybeInfos = tsUtils.parseDataForExtendsClassCompletion(sourceFile, position)
+    if (!maybeInfos) return []
+    const { accessedObject, className, replacementSpan } = maybeInfos
 
     // first, given the position, we go back
-    const contextName = yield* Nano.option(
-      AST.findImportedModuleIdentifierByPackageAndNameOrBarrel(
-        sourceFile,
-        "effect",
-        "Context"
-      )
-    )
-    const contextIdentifier = Option.match(contextName, {
-      onNone: () => "Context",
-      onSome: (_) => _.text
-    })
+    const contextIdentifier = tsUtils.findImportedModuleIdentifierByPackageAndNameOrBarrel(
+      sourceFile,
+      "effect",
+      "Context"
+    ) || "Context"
 
     // ensure accessed is an identifier
     if (contextIdentifier !== accessedObject.text) return []
