+++ src/core/TypeScriptApi.ts
@@ -1,6 +1,3 @@
-import * as Array from "effect/Array"
-import { pipe } from "effect/Function"
-import { hasProperty, isFunction, isObject, isString } from "effect/Predicate"
 import type ts from "typescript"
 import * as Nano from "../core/Nano.js"
 
@@ -161,113 +158,3 @@ export interface TypeScriptProgram extends _TypeScriptProgram {}
 export const TypeScriptProgram = Nano.Tag<TypeScriptProgram>("TypeScriptProgram")
 
 export const ChangeTracker = Nano.Tag<ts.textChanges.ChangeTracker>("ChangeTracker")
-
-export interface ModuleWithPackageInfo {
-  name: string
-  version: string
-  hasEffectInPeerDependencies: boolean
-  contents: any
-  packageDirectory: string
-  referencedPackages: Array<string>
-  exportsKeys: Array<string>
-}
-
-export function parsePackageContentNameAndVersionFromScope(v: unknown): ModuleWithPackageInfo | undefined {
-  if (!isObject(v)) return
-  if (!hasProperty(v, "packageJsonScope")) return
-  if (!v.packageJsonScope) return
-  const packageJsonScope = v.packageJsonScope
-  if (!hasProperty(packageJsonScope, "contents")) return
-  if (!hasProperty(packageJsonScope.contents, "packageJsonContent")) return
-  const packageJsonContent = packageJsonScope.contents.packageJsonContent
-  if (!hasProperty(packageJsonContent, "name")) return
-  if (!hasProperty(packageJsonContent, "version")) return
-  if (!hasProperty(packageJsonScope, "packageDirectory")) return
-  if (!isString(packageJsonScope.packageDirectory)) return
-  const { name, version } = packageJsonContent
-  if (!isString(name)) return
-  if (!isString(version)) return
-  const hasEffectInPeerDependencies = hasProperty(packageJsonContent, "peerDependencies") &&
-    isObject(packageJsonContent.peerDependencies) &&
-    hasProperty(packageJsonContent.peerDependencies, "effect")
-
-  const referencedPackages = Object.keys({
-    ...(hasProperty(packageJsonContent, "dependencies") && isObject(packageJsonContent.dependencies)
-      ? packageJsonContent.dependencies
-      : {}),
-    ...(hasProperty(packageJsonContent, "peerDependencies") && isObject(packageJsonContent.peerDependencies)
-      ? packageJsonContent.peerDependencies
-      : {}),
-    ...(hasProperty(packageJsonContent, "devDependencies") && isObject(packageJsonContent.devDependencies)
-      ? packageJsonContent.devDependencies
-      : {})
-  })
-
-  const exportsKeys = Object.keys(
-    hasProperty(packageJsonContent, "exports") && isObject(packageJsonContent.exports)
-      ? packageJsonContent.exports
-      : {}
-  )
-
-  return {
-    name: name.toLowerCase(),
-    version: version.toLowerCase(),
-    hasEffectInPeerDependencies,
-    contents: packageJsonContent,
-    packageDirectory: packageJsonScope.packageDirectory,
-    referencedPackages,
-    exportsKeys
-  }
-}
-
-export const resolveModulePattern = Nano.fn("resolveModulePattern")(
-  function*(sourceFile: ts.SourceFile, pattern: string) {
-    if (pattern.indexOf("*") === -1) return [pattern.toLowerCase()]
-    const ts = yield* Nano.service(TypeScriptApi)
-    const packageJsonScope = parsePackageContentNameAndVersionFromScope(sourceFile)
-    const referencedPackages: Array<string> = []
-    for (const statement of sourceFile.statements) {
-      if (ts.isImportDeclaration(statement) && ts.isStringLiteral(statement.moduleSpecifier)) {
-        const moduleSpecifier = statement.moduleSpecifier.text.toLowerCase()
-        const packageName = moduleSpecifier.startsWith("@")
-          ? moduleSpecifier.split("/", 2).join("/")
-          : moduleSpecifier.split("/", 1).join("/")
-        referencedPackages.push(packageName)
-      }
-    }
-    return pipe(
-      referencedPackages.concat(packageJsonScope?.referencedPackages || []),
-      Array.dedupe,
-      Array.map((packageName) => packageName.toLowerCase()),
-      Array.filter((packageName) =>
-        pattern.endsWith("*") &&
-        packageName.startsWith(pattern.toLowerCase().substring(0, pattern.length - 1))
-      )
-    )
-  }
-)
-
-export function makeGetModuleSpecifier(ts: TypeScriptApi) {
-  if (
-    !(hasProperty(ts, "moduleSpecifiers") && hasProperty(ts.moduleSpecifiers, "getModuleSpecifier") &&
-      isFunction(ts.moduleSpecifiers.getModuleSpecifier))
-  ) return
-  const _internal = ts.moduleSpecifiers.getModuleSpecifier
-  return (
-    compilerOptions: ts.CompilerOptions,
-    importingSourceFile: ts.SourceFile,
-    importingSourceFileName: string,
-    toFileName: string,
-    host: any,
-    options?: any
-  ): string => {
-    return _internal(
-      compilerOptions,
-      importingSourceFile,
-      importingSourceFileName,
-      toFileName,
-      host,
-      options
-    )
-  }
-}
