+++ src/goto.ts
@@ -2,6 +2,7 @@ import type * as ts from "typescript"
 import type * as Nano from "./core/Nano"
 import type * as TypeCheckerApi from "./core/TypeCheckerApi"
 import type * as TypeScriptApi from "./core/TypeScriptApi"
+import type * as TypeScriptUtils from "./core/TypeScriptUtils"
 import { effectRpcDefinition } from "./goto/effectRpcDefinition"
 
 export function goto(
@@ -11,7 +12,10 @@ export function goto(
 ): Nano.Nano<
   ts.DefinitionInfoAndBoundSpan | undefined,
   never,
-  TypeScriptApi.TypeScriptApi | TypeScriptApi.TypeScriptProgram | TypeCheckerApi.TypeCheckerApi
+  | TypeScriptApi.TypeScriptApi
+  | TypeScriptUtils.TypeScriptUtils
+  | TypeScriptApi.TypeScriptProgram
+  | TypeCheckerApi.TypeCheckerApi
 > {
   return effectRpcDefinition(applicableGotoDefinition, sourceFile, position)
 }
