+++ src/refactors/removeUnnecessaryEffectGen.ts
@@ -1,9 +1,9 @@
 import * as Option from "effect/Option"
-import * as AST from "../core/AST.js"
 import * as LSP from "../core/LSP.js"
 import * as Nano from "../core/Nano.js"
 import * as TypeParser from "../core/TypeParser.js"
 import * as TypeScriptApi from "../core/TypeScriptApi.js"
+import * as TypeScriptUtils from "../core/TypeScriptUtils.js"
 
 /**
  * Refactor to remove unnecessary `Effect.gen` calls.
@@ -38,10 +38,11 @@ export const removeUnnecessaryEffectGen = LSP.createRefactor({
   name: "removeUnnecessaryEffectGen",
   description: "Remove unnecessary Effect.gen",
   apply: Nano.fn("removeUnnecessaryEffectGen.apply")(function*(sourceFile, textRange) {
+    const tsUtils = yield* Nano.service(TypeScriptUtils.TypeScriptUtils)
     const typeParser = yield* Nano.service(TypeParser.TypeParser)
 
     for (
-      const nodeToReplace of yield* AST.getAncestorNodesInRange(sourceFile, textRange)
+      const nodeToReplace of tsUtils.getAncestorNodesInRange(sourceFile, textRange)
     ) {
       const maybeNode = yield* Nano.option(typeParser.unnecessaryEffectGen(nodeToReplace))
