+++ src/refactors/toggleTypeAnnotation.ts
@@ -1,25 +1,26 @@
-import * as ReadonlyArray from "effect/Array"
+import * as Array from "effect/Array"
 import { pipe } from "effect/Function"
 import * as Option from "effect/Option"
-import * as AST from "../core/AST.js"
 import * as LSP from "../core/LSP.js"
 import * as Nano from "../core/Nano.js"
 import * as TypeCheckerApi from "../core/TypeCheckerApi.js"
 import * as TypeScriptApi from "../core/TypeScriptApi.js"
+import * as TypeScriptUtils from "../core/TypeScriptUtils.js"
 
 export const toggleTypeAnnotation = LSP.createRefactor({
   name: "toggleTypeAnnotation",
   description: "Toggle type annotation",
   apply: Nano.fn("toggleTypeAnnotation.apply")(function*(sourceFile, textRange) {
     const ts = yield* Nano.service(TypeScriptApi.TypeScriptApi)
+    const tsUtils = yield* Nano.service(TypeScriptUtils.TypeScriptUtils)
     const typeChecker = yield* Nano.service(TypeCheckerApi.TypeCheckerApi)
 
     const maybeNode = pipe(
-      yield* AST.getAncestorNodesInRange(sourceFile, textRange),
-      ReadonlyArray.filter((node) => ts.isVariableDeclaration(node) || ts.isPropertyDeclaration(node)),
-      ReadonlyArray.filter((node) => AST.isNodeInRange(textRange)(node.name)),
-      ReadonlyArray.filter((node) => !!node.initializer),
-      ReadonlyArray.head
+      tsUtils.getAncestorNodesInRange(sourceFile, textRange),
+      Array.filter((node) => ts.isVariableDeclaration(node) || ts.isPropertyDeclaration(node)),
+      Array.filter((node) => tsUtils.isNodeInRange(textRange)(node.name)),
+      Array.filter((node) => !!node.initializer),
+      Array.head
     )
 
     if (Option.isNone(maybeNode)) return yield* Nano.fail(new LSP.RefactorNotApplicableError())
@@ -57,7 +58,7 @@ export const toggleTypeAnnotation = LSP.createRefactor({
             changeTracker.insertNodeAt(
               sourceFile,
               node.name.end,
-              yield* AST.simplifyTypeNode(initializerTypeNode),
+              tsUtils.simplifyTypeNode(initializerTypeNode),
               {
                 prefix: ": "
               }
