+++ src/refactors/typeToEffectSchemaClass.ts
@@ -4,6 +4,7 @@ import * as LSP from "../core/LSP.js"
 import * as Nano from "../core/Nano.js"
 import * as TypeCheckerApi from "../core/TypeCheckerApi.js"
 import * as TypeScriptApi from "../core/TypeScriptApi.js"
+import * as TypeScriptUtils from "../core/TypeScriptUtils.js"
 import * as SchemaGen from "../utils/SchemaGen.js"
 
 export const typeToEffectSchemaClass = LSP.createRefactor({
@@ -11,6 +12,7 @@ export const typeToEffectSchemaClass = LSP.createRefactor({
   description: "Refactor to Schema.Class",
   apply: Nano.fn("typeToEffectSchemaClass.apply")(function*(sourceFile, textRange) {
     const ts = yield* Nano.service(TypeScriptApi.TypeScriptApi)
+    const tsUtils = yield* Nano.service(TypeScriptUtils.TypeScriptUtils)
     const typeChecker = yield* Nano.service(TypeCheckerApi.TypeCheckerApi)
 
     const maybeNode = yield* SchemaGen.findNodeToProcess(sourceFile, textRange)
@@ -24,6 +26,7 @@ export const typeToEffectSchemaClass = LSP.createRefactor({
       apply: pipe(
         SchemaGen.applyAtNode(sourceFile, node, true),
         Nano.provideService(TypeCheckerApi.TypeCheckerApi, typeChecker),
+        Nano.provideService(TypeScriptUtils.TypeScriptUtils, tsUtils),
         Nano.provideService(TypeScriptApi.TypeScriptApi, ts)
       )
     })
