+++ src/refactors/wrapWithEffectGen.ts
@@ -2,18 +2,19 @@ import * as ReadonlyArray from "effect/Array"
 import { pipe } from "effect/Function"
 import * as Option from "effect/Option"
 import type ts from "typescript"
-import * as AST from "../core/AST.js"
 import * as LSP from "../core/LSP.js"
 import * as Nano from "../core/Nano.js"
 import * as TypeCheckerApi from "../core/TypeCheckerApi.js"
 import * as TypeParser from "../core/TypeParser.js"
 import * as TypeScriptApi from "../core/TypeScriptApi.js"
+import * as TypeScriptUtils from "../core/TypeScriptUtils.js"
 
 export const wrapWithEffectGen = LSP.createRefactor({
   name: "wrapWithEffectGen",
   description: "Wrap with Effect.gen",
   apply: Nano.fn("wrapWithEffectGen.apply")(function*(sourceFile, textRange) {
     const ts = yield* Nano.service(TypeScriptApi.TypeScriptApi)
+    const tsUtils = yield* Nano.service(TypeScriptUtils.TypeScriptUtils)
     const typeChecker = yield* Nano.service(TypeCheckerApi.TypeCheckerApi)
     const typeParser = yield* Nano.service(TypeParser.TypeParser)
 
@@ -34,30 +35,18 @@ export const wrapWithEffectGen = LSP.createRefactor({
     )
 
     const maybeNode = yield* pipe(
-      yield* AST.getAncestorNodesInRange(sourceFile, textRange),
+      tsUtils.getAncestorNodesInRange(sourceFile, textRange),
       ReadonlyArray.map(findEffectToWrap),
       Nano.firstSuccessOf,
       Nano.option
     )
     if (Option.isNone(maybeNode)) return yield* Nano.fail(new LSP.RefactorNotApplicableError())
 
-    const effectModuleIdentifier = Option.match(
-      yield* Nano.option(
-        AST.findImportedModuleIdentifier(
-          sourceFile,
-          (node) =>
-            pipe(
-              typeParser.importedEffectModule(node),
-              Nano.option,
-              Nano.map(Option.isSome)
-            )
-        )
-      ),
-      {
-        onNone: () => "Effect",
-        onSome: (node) => node.text
-      }
-    )
+    const effectModuleIdentifier = tsUtils.findImportedModuleIdentifierByPackageAndNameOrBarrel(
+      sourceFile,
+      "effect",
+      "Effect"
+    ) || "Effect"
 
     return {
       kind: "refactor.rewrite.effect.wrapWithEffectGen",
@@ -66,9 +55,9 @@ export const wrapWithEffectGen = LSP.createRefactor({
         Nano.gen(function*() {
           const changeTracker = yield* Nano.service(TypeScriptApi.ChangeTracker)
 
-          const effectGen = yield* AST.createEffectGenCallExpressionWithBlock(
+          const effectGen = tsUtils.createEffectGenCallExpressionWithBlock(
             effectModuleIdentifier,
-            yield* AST.createReturnYieldStarStatement(maybeNode.value)
+            tsUtils.createReturnYieldStarStatement(maybeNode.value)
           )
 
           changeTracker.replaceNode(sourceFile, maybeNode.value, effectGen)
