+++ src/transform.ts
@@ -9,6 +9,7 @@ import * as Nano from "./core/Nano"
 import * as TypeCheckerApi from "./core/TypeCheckerApi"
 import * as TypeParser from "./core/TypeParser"
 import * as TypeScriptApi from "./core/TypeScriptApi"
+import * as TypeScriptUtils from "./core/TypeScriptUtils"
 import { diagnostics } from "./diagnostics"
 
 export default function(
@@ -21,14 +22,11 @@ export default function(
       // run the diagnostics and pipe them into addDiagnostic
       pipe(
         LSP.getSemanticDiagnosticsWithCodeFixes(diagnostics, sourceFile),
-        Nano.provideService(TypeScriptApi.TypeScriptApi, tsInstance),
-        Nano.provideService(TypeScriptApi.TypeScriptProgram, program),
+        TypeParser.nanoLayer,
+        TypeScriptUtils.nanoLayer,
         Nano.provideService(TypeCheckerApi.TypeCheckerApi, program.getTypeChecker()),
-        Nano.provideService(TypeParser.TypeParser, TypeParser.make(tsInstance, program.getTypeChecker())),
-        Nano.provideService(
-          TypeCheckerApi.TypeCheckerApiCache,
-          TypeCheckerApi.makeTypeCheckerApiCache()
-        ),
+        Nano.provideService(TypeScriptApi.TypeScriptProgram, program),
+        Nano.provideService(TypeScriptApi.TypeScriptApi, tsInstance),
         Nano.provideService(
           LanguageServicePluginOptions.LanguageServicePluginOptions,
           LanguageServicePluginOptions.parse(pluginConfig)
