+++ test/perf.ts
@@ -9,6 +9,7 @@ import * as Nano from "../src/core/Nano"
 import * as TypeCheckerApi from "../src/core/TypeCheckerApi"
 import * as TypeParser from "../src/core/TypeParser"
 import * as TypeScriptApi from "../src/core/TypeScriptApi"
+import * as TypeScriptUtils from "../src/core/TypeScriptUtils"
 import { diagnostics } from "../src/diagnostics"
 import { createServicesWithMockedVFS } from "./utils/mocks.js"
 
@@ -22,7 +23,6 @@ function testAllDagnostics() {
   console.log("running a couple of times")
   const totalSamples = 1000
   let totalTime = 0
-  const cache = TypeCheckerApi.makeTypeCheckerApiCache()
   for (const exampleFileName of allExampleFiles) {
     const sourceText = fs.readFileSync(path.join(getExamplesDiagnosticsDir(), exampleFileName))
       .toString("utf8")
@@ -31,14 +31,11 @@ function testAllDagnostics() {
       const start = performance.now()
       pipe(
         LSP.getSemanticDiagnosticsWithCodeFixes(diagnostics, example.sourceFile),
-        Nano.provideService(TypeScriptApi.TypeScriptApi, ts),
-        Nano.provideService(TypeScriptApi.TypeScriptProgram, example.program),
+        TypeParser.nanoLayer,
         Nano.provideService(TypeCheckerApi.TypeCheckerApi, example.program.getTypeChecker()),
-        Nano.provideService(
-          TypeCheckerApi.TypeCheckerApiCache,
-          cache
-        ),
-        Nano.provideService(TypeParser.TypeParser, TypeParser.make(ts, example.program.getTypeChecker())),
+        Nano.provideService(TypeScriptApi.TypeScriptProgram, example.program),
+        TypeScriptUtils.nanoLayer,
+        Nano.provideService(TypeScriptApi.TypeScriptApi, ts),
         Nano.provideService(LanguageServicePluginOptions.LanguageServicePluginOptions, {
           diagnostics: true,
           diagnosticSeverity: {},
